better normal check for varscan

+++ 06d_call_SNVs_and_indels.varscan.sh
@@ -136,12 +136,16 @@ fi
 # check if command finished
 if [[ "$check_finish" == 0 ]]; then
     # log to main
-    echo "02: ${tumor}__${normal} VarScan2 variant calling completed." | tee -a main.log
+    echo "06: ${tumor}__${normal} VarScan2 variant calling completed." | tee -a main.log
     # move logfile
     mv ${tumor}__${normal}.VarScan.log all_logfiles
     # delete pileups
     rm varscan/pileups/${tumor}.pileup
     if [[ -e varscan/pileups/${normal}.pileup ]]; then
-        rm varscan/pileups/${normal}.pileup
+        # how many T are paired with N
+        paired_tumors=$(grep -c ",${normal}$" tumors_and_normals.csv)
+        if [[ $(ls *__${normal}.varscan.all.Somatic.hc.wes.vcf.gz | wc -l) == "${paired_tumors}" ]]; then
+            rm varscan/pileups/${normal}.pileup
+        fi
     fi
 fi
